Fix certificate paths to use perl2host
Moving the certificates into a truly temporary path for the duration
of the tests broke msys as the absolute path wasn't adapted for the
msys platform.  Ensure to convert and expand the path with perl2host
before copying and passing in the connection string.

Discussion: https://postgr.es/m/YacT3tm97xziSUFw@paquier.xyz