Refactor SSL tests
The SSL support in libpq, pg_cryptohash and other relevant parts of the
codebase, has been abstracted such that support for alternative SSL
libraries can be implemented. The tests were however quite heavily tied
to OpenSSL, and how the OpenSSL implementation sets up the server. This
refactors the SSLServer.pm module into SSL::Server which abstracts SSL
library specific setup, which in turn is implemented in SSL::Backend::X
for each SSL library. Rudimentary POD documentation is also added.

The main difference for testcode is that switch_server_cert() now takes
a set of parameters where those which are library specific are passed
to the implementing module. Resolving the path to the sslkey is also
moved to the backend module, and is now a function call rather than a
hash lookup.

While not strictly related to the refactoring, the preplanning of test
cases is removed in favor of calling done_testing(), as the number is
quite hard to calculate by hand given that connection tests perform
subtests.

This was extracted from a larger patchset on SSL library support, as of
now only OpenSSL is supported however.

Discussion: https://postgr.es/m/<tbd>
Discussion: https://postgr.es/m/FAB21FC8-0F62-434F-AA78-6BD9336D630A@yesql.se