Set fixed dates for test certificates validity
Rather than specifying a validity of 10 000 days into the future
for test certificate generation, this hardcodes the notBefore and
notAfter attributes to set values. This will allow writing tests
which will not depend on when the certificates were generated. An
upcoming patch for adding notBefore and notAfter to pg_stat_ssl
and sslinfo requires this.

Discussion: https://postgr.es/m/182b8565486.10af1a86f158715.2387262617218380588@highgo.ca