Speed up pg_regress server testing
Instead of connecting to the server with psql to check if it is ready
for running tests, pg_regress now use PQPing which avoids performing
expensive system() calls on Windows.

The frequency of tests is also increased in order to connect to the
server faster.

This patch is part of a larger effort to make testing consume fewer
resources in order to be able to fit more tests into the available
CI constraints.

Discussion: https://postgr.es/m/20230823192239.jxew5s3sjru63lio@awork3.anarazel.de