XXX fix libcurl link error
The ftp/curl port appears to be missing a minimum version dependency on
libssh2, so the following starts showing up after upgrading to curl
8.11.1_1:

    libcurl.so.4: Undefined symbol "libssh2_session_callback_set2"

But 13.3 is EOL, so it's not clear if anyone would be interested in a
bug report, and a FreeBSD 14 Cirrus image is in progress. Hack past it
for now.