Report test failure rather than aborting in case of timeout
When an querying an interactive, or background, psql process the
call would die() in case of a timeout since it used the IPC::Run
timeout construct and not a timer. This aborts tests prematurely
which makes debugging harder, and stops getting results for the
remaining tests in the suite.  This converts to using a timer
object instead which will transfer control back to the caller
regardless, returning undef in case of a timeout along with
logging a test failure. Affected callsites are also updated to
reflect this.

Reviewed-by: Heikki Linnakangas <hlinnaka@iki.fi>
Reviewed-by: Andrew Dunstan <andrew@dunslane.net>
Discussion: https://postgr.es/m/1100715.1712265845@sss.pgh.pa.us