Handle --help and --version as normal commandline parameters
Traditionally all postgres apps has treated --help, -?, --version and
-V as special case options which have to be the first parameter in the
commandline, else they are considered unrecognized.  This is not the
typical way to handle parameters in command line utilities, leading to
a bad user experience when a documented parameter is erroring out with
an unrecognized error.

This moves --help, -?, --version and -V into the common command line
parameter processing allowing them to be used at any argument position.
It turned out that our apps had very varied support for these in the
long_options array and getopt_long call.  Some apps had the long option
defined but not the short option, some had both and some had neither of
them.  Handling the actual parameters was also done in different ways,
some used inline code and others the handle_help_version_opts helper.
psql was its own case which had both, so removing the position one code
was all that was needed.  The getopt_long handling in psql is now the
one place which is a little different from the others as it's left the
way it was given that it already worked.

Since getopt_long will return '?' for invalid parameters as well as for
the -? parameter without any way to distinguish the case at hand, there
is a special function added which is used to separate the two.

While there, ensure that progname is used consistently when printing
the name of the app, and move progname from the global namespace to
improve consistency across the code.  TODO: this part of the patch
should probably be either ripped out or at least separated into its
own commit.

Author: Daniel Gustafsson <daniel@yesql.se>
Reviewed-by: ...
Reported-by: Matthias Hörmann <matthias.hoermann@saltation.com>
Discussion: https://postgr.es/m/19042-347869ae37c993f4@postgresql.org